New update
done some more work to implement boss fights, i could do it the same way i did a normal npc fight but that would be inefficient, i am trying to figure out a way to make stuff like the whole fight work with one code, and it just alternates depending on whether you are fighting a boss or a mini fight. I have a couple of ideas that include interfaces and make boss and enemy implement it, but we'll see about that in the near future.

+++ src/Boss.java
@@ -1,22 +1,22 @@
 public class Boss {
-    int health;
-    int armour;
-    int damage;
-    public Boss (int damagePassed, int healthPassed, int armourPassed) {
+    double health;
+    double armour;
+    double damage;
+    public Boss (double damagePassed, double healthPassed, double armourPassed) {
         health = healthPassed;
         armour = armourPassed;
         damage = damagePassed;
     }
 
-    public int getHealth () {
+    public double getHealth () {
         return health;
     }
 
-    public int getArmour () {
+    public double getArmour () {
         return armour;
     }
 
-    public int getDamage () {
+    public double getDamage () {
         return damage;
     }
 }
